Smaller refactors
